+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.4.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/inquirer-common@0.7.0
+    -   @sap-ux/project-access@1.30.0
+
+## 0.3.104
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.43
+
 ## 0.3.103
 
 ### Patch Changes